% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dissimilarity-based.R
\name{add_metadata_to_dm_clmns}
\alias{add_metadata_to_dm_clmns}
\title{Add metadata to an additional column in column formatted
 dissimilarities dataframe}
\usage{
add_metadata_to_dm_clmns(dm_clmns, metadata_map, cat)
}
\arguments{
\item{dm_clmns}{The dissimilarities dataframe produced using \code{
convert_dm_to_3_column()}.}

\item{metadata_map}{The metadata dataframe.}

\item{cat}{The header string from the metadata map corresponding to the data
you would like to add to the dissimilarities dataframe.}
}
\description{
Add metadata to an additional column in column formatted
 dissimilarities dataframe
}

