% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/routine_analysis_functions.R
\name{calc_diversity}
\alias{calc_diversity}
\title{Calculate diversity values}
\usage{
calc_diversity(taxon_table, metric)
}
\arguments{
\item{taxon_table}{The taxon table}

\item{metric}{The diversity metric to calculate}
}
\description{
Calculate diversity values from a taxon table
}

