% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dissimilarity-based.R
\name{calc_dm}
\alias{calc_dm}
\title{Calculate a dissimilarity matrix from a taxa table}
\usage{
calc_dm(tax_table)
}
\arguments{
\item{tax_table}{The taxa table.}
}
\value{
A variable of class 'dist'.
}
\description{
Currently calculates Bray-Curtis dissimilarities after
 performing square-root transformations on the data. Soon, other metrics
 will be available.
}

