% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dissimilarity-based.R
\name{calc_ordination}
\alias{calc_ordination}
\title{Calculate Point Coordinates in an Ordination}
\usage{
calc_ordination(dm, ord_type, metadata_map, constrain_factor)
}
\arguments{
\item{dm}{Dissimilarity matrix.}

\item{ord_type}{The type of ordination. 'NMDS', 'PCoA', or 'constrained' are 
the current accepted values.}

\item{metadata_map}{Required if 'constrained' ord_type.}

\item{constrain_factor}{Required if 'constrained' ord_type.}
}
\value{
A data frame consisting of the coordinates.
}
\description{
Use to generate ordination axis scores to use when plotting
  ordination.
}
\details{
Impliments nonmetric multidimensional scaling, principal coordinates
  analysis, or constrained ordination (via \code{\link{capscale}} in
  \code{\link{vegan}}).
}
\examples{
dm = calc_dm(fruits_veggies$data_loaded)
ord = calc_ordination(dm, "nmds")
}
\concept{
Dissimilarity calculation and manipulation
}

