% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dissimilarity-based.R
\name{calc_pairwise_permanovas}
\alias{calc_pairwise_permanovas}
\title{Calculate pairwise PERMANOVA results}
\usage{
calc_pairwise_permanovas(dm, metadata_map, compare_header)
}
\arguments{
\item{dm}{Dissimilarity matrix of class \code{dist}.}

\item{metadata_map}{The metadata mapping dataframe with samples matching and 
in the same order as the ones in the provided dm.}

\item{compare_header}{The header in the metadata mapping dataframe with the 
factor levels to use for the pairwise comparisons.}
}
\value{
A dataframe with R2 and P values.
}
\description{
The \code{\link[vegan]{adonis}} function in the 
  \code{\link[vegan]{vegan}} package does not provide a way to calculate 
  pairwise post-hoc comparisons between factor levels. This function 
  calculates PERMANOVA results using the \code{\link[vegan]{adonis}}
  function for all factor level pairs. Raw p values are returned, but it is
  recommended to use the provided FDR corrected p values since the multiple
  comparisons can raise your likelihood of false positive differences.
}
\examples{
dm = calc_dm(fruits_veggies$data_loaded)
calc_pairwise_permanovas(dm, fruits_veggies$map_loaded, "Sample_type")
}
\concept{
Dissimilarity calculation and manipulation
}
