% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/misc_analyses.R
\name{calc_prop_taxa_from_sample_type}
\alias{calc_prop_taxa_from_sample_type}
\title{Calculate the proportion of taxa in a set of samples that are
 also observed in another sample type}
\usage{
calc_prop_taxa_from_sample_type(input, type_header, primary_type, source_type,
  within_cat)
}
\arguments{
\item{input}{Data from load_taxa_table()}

\item{type_header}{The header label from the metadata map that has the
labels which indicate the types of samples you want to use.}

\item{primary_type}{The sample type for which you would like to calculate
the proportion of taxa that are also observed in the source_type.}

\item{source_type}{The sample type to use as the source type.}

\item{within_cat}{(Optional) Specify a header label from the metadata map
that tells the function to restrict pairs to within levels of this factor.}
}
\description{
Calculate the proportion of taxa in a set of samples that are
 also observed in another sample type
}

