% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/taxonomy-based.R
\name{calc_taxa_changes}
\alias{calc_taxa_changes}
\title{Calculate the changes in taxonomic relative abundances compared to
 controls}
\usage{
calc_taxa_changes(ts, metadata_map, block_header, treatment_header,
  control_label)
}
\arguments{
\item{ts}{A taxa summary dataframe as generated by
\code{summarize_taxonomy()}.}

\item{metadata_map}{A metadata mapping dataframe. Usually,
\code{input$map_loaded}.}

\item{block_header}{The header name from the metadata map that contains the
block (i.e. sample group) designations.}

\item{treatment_header}{The header name from the metadata map that contains
the treatment designations including which samples are controls.}

\item{control_label}{The designation within the \code{treatment_header}
column that corresponds to the control samples.}
}
\description{
Calculate the change relative to controls within sample groups
 (blocks) on a per taxon basis.
}

