% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/taxa_by_sample_type_functions.R
\name{calc_taxa_means}
\alias{calc_taxa_means}
\title{Calculate mean taxa values across a specified factor}
\usage{
calc_taxa_means(taxa_table, metadata_map, summarize_by_factor,
  return_map = TRUE)
}
\arguments{
\item{taxa_table}{The taxa table.}

\item{metadata_map}{The metadata mapping data frame.}

\item{summarize_by_factor}{Category in mapping file to summarize by.}

\item{return_map}{Whether or not to return summarized mapping files. If true,
will return a list (default: TRUE).}
}
\value{
If \code{return_map == TRUE}, returns a list with a taxon table
 (data_loaded) and a mapping data frame (map_loaded). Otherwise, returns a
 taxon table of class data frame.
}
\description{
Calculate mean taxa values across a specified factor
}

