% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/taxonomy-based.R
\name{collapse_taxonomy}
\alias{collapse_taxonomy}
\title{Collapse taxonomy dataframe to character vector}
\usage{
collapse_taxonomy(taxonomy_df)
}
\arguments{
\item{taxonomy_df}{The dataframe containing taxonomy as loaded by \code{
load_taxa_table()}.}
}
\description{
Collapse taxonomy dataframe to character vector
}

