% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dissimilarity-based.R
\name{convert_dm_to_3_column}
\alias{convert_dm_to_3_column}
\title{Convert dissimilarity matrix to 3 column format}
\usage{
convert_dm_to_3_column(dm)
}
\arguments{
\item{dm}{Dissimilarity matrix of either class 'dist' or class 'data.frame'}
}
\description{
This is useful for performing analyses on dissimilarity values
}

