% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/routine_analysis_functions.R
\name{convert_to_relative_abundances}
\alias{convert_to_relative_abundances}
\title{convert_to_relative_abundances}
\usage{
convert_to_relative_abundances(input)
}
\arguments{
\item{input}{Either a dataset as generated using 'load_taxon_table' which
includes a mapping file or an individual taxon table}
}
\description{
Convert taxon table or taxon table in data input to relative
abundances
}

