% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/input_output_filter.R
\name{export_taxa_table}
\alias{export_taxa_table}
\title{Export a taxa table as a text file}
\usage{
export_taxa_table(input, out_fp, map_fp)
}
\arguments{
\item{input}{The input dataset as loaded by \code{load_taxa_table()} or
a taxa table of class \code{data.frame}.}

\item{out_fp}{The output filepath.}

\item{map_fp}{[OPTIONAL] The metadata map output filepath if you want to
write it to file.}
}
\description{
A convenient way to export a loaded taxa table as a text file.
 Taxonomy strings will appear in the right most column. This is also a good
 way to save an taxa table to be loaded later. Output format is tab-delimited.
}
\examples{
\dontrun{
export_taxa_table(fruits_veggies, "exported_taxa_table", "exported_map_file")
}
}

