% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_output_filter.R
\name{filter_data}
\alias{filter_data}
\title{Filter samples from dataset}
\usage{
filter_data(input, filter_cat, filter_vals, keep_vals)
}
\arguments{
\item{input}{The input dataset as loaded by \code{\link{load_taxa_table}}.}

\item{filter_cat}{The map_fp header string for the factor you would like to
use to filter samples.}

\item{filter_vals}{The values within the filter category (vector or single
value) you would like to use to remove samples from the imported data.}

\item{keep_vals}{Alternatively, keep only samples represented by these
values.}
}
\value{
A list variable with (1) the loaded taxa table, (2) the loaded
  mapping file, and optionally (3) the loaded taxonomy information.
}
\description{
Filter out or keep particular samples in a dataset based on
  contextual metadata.
}
\examples{
\dontrun{
ex_in_filt = filter_data(input = "example_input", filter_cat = "Sample_type",
                         filter_vals = c("mushrooms", "strawberries"))
}
}
\concept{Taxa table manipulation}
