% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/input_output_filter.R
\name{filter_samples_by_counts}
\alias{filter_samples_by_counts}
\title{Filter Samples from a Dataset based on number of sequences}
\usage{
filter_samples_by_counts(input, min_seqs)
}
\arguments{
\item{input}{The input dataset as loaded by \code{load_taxa_table()}.}

\item{min_seqs}{A sample must have at minimum this number of sequences
in order to be retained.}
}
\value{
A list variable with (1) the loaded taxa table, (2)
 the loaded mapping file, and optionally (3) the loaded taxonomy information.
}
\description{
Filter Samples from a Dataset based on number of sequences
}

