% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/taxonomy-based.R
\name{filter_taxa_from_input}
\alias{filter_taxa_from_input}
\title{Filter taxa from a loaded dataset}
\usage{
filter_taxa_from_input(input, filter_thresh, taxa_to_keep, taxa_to_remove,
  at_spec_level, taxa_IDs_to_keep, taxa_IDs_to_remove)
}
\arguments{
\item{input}{Input data (a list variable) from \code{load_taxa_table()}.}

\item{filter_thresh}{Filter OTUs less than this number based on mean OTU
table values.}

\item{taxa_to_keep}{Keep only taxa that contain these names. Vector or string.}

\item{taxa_to_remove}{Remove taxa that contain these names. Vector or string.}

\item{at_spec_level}{If included, only keep/remove matches at this specific
taxonomy level(s) (a number/numbers referring to the taxonomy
column(s)).}
}
\description{
Filter taxa from a loaded dataset
}
\details{
Can use one or more of the parameters to do filtering. Threshold
         filtering takes precidence over taxa filtering. If taxa to keep and
         taxa to remove are both included, taxa to remove will be
         removed if the parameter entries conflict. Taxa are found using
         grep, so it is not necessary to use the entire names of taxa as long
         as they are not ambiguous. Can filter using taxa IDs (i.e. row
         names in taxa table).
}

