% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_documentation.R
\docType{data}
\name{fruits_veggies}
\alias{fruits_veggies}
\title{Fruits and Vegetables Bacterial Community Data}
\format{
A list with 3 elements.
}
\source{
Adapted from \url{
 http://journals.plos.org/plosone/article?id=10.1371/journal.pone.0059310}
}
\usage{
fruits_veggies
}
\description{
Taxa table, metadata map, and bacterial taxonomy. Produced using \code{
mctoolsr::load_taxa_table()}
}
\keyword{datasets}
