% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets_documentation.R
\docType{data}
\name{fruits_veggies_OTUs_tree}
\alias{fruits_veggies_OTUs_tree}
\title{Fruits and Vegetables Bacterial Phylotype Tree}
\format{phylo}
\source{
Adapted from \url{
 http://journals.plos.org/plosone/article?id=10.1371/journal.pone.0059310}
}
\usage{
fruits_veggies_OTUs_tree
}
\description{
Phylogenetic tree created using pynast and Fasttree
}
\keyword{datasets}

