% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_output_filter.R
\name{load_2_dms}
\alias{load_2_dms}
\title{Load two dissimilarity matrices for use with mctoolsr}
\usage{
load_2_dms(dm1_fp, dm2_fp, map_fp, filter_cat, filter_vals, keep_vals)
}
\arguments{
\item{dm1_fp}{Dissimilarity matrix filepath (tab-delimited text).}

\item{dm2_fp}{Dissimilarity matrix filepath (tab-delimited text).}

\item{map_fp}{Metadata mapping filepath.}

\item{filter_cat}{[OPTIONAL] The map_fp header string for the factor you
would like to use to filter samples.}

\item{filter_vals}{[OPTIONAL] The values within the filter category (vector
or single value) you would like to use to remove samples from the imported
data.}

\item{keep_vals}{[OPTIONAL] Alternatively, keep only samples represented by
these values.}
}
\value{
A list variable with (1) the loaded dissimilarity matrix 1, (2) the 
  loaded dissimilarity matrix 2, and (3) the loaded mapping file.
}
\description{
Load in two dissimilarity matrices and a corresponding metadata 
  mapping file. Useful for Mantel tests when dms are already generated.
}
\examples{
\dontrun{
load_dm("filepath_to_dissim_matrix1.txt", "filepath_to_dissim_matrix1.txt",
  "filepath_to_mapping_file.txt", "sample_type", filter_vals = "blank")
}
}
\concept{
Load external data
}
