% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_output_filter.R
\name{load_dm}
\alias{load_dm}
\title{Load a dissimilarity matrix for use with mctoolsr}
\usage{
load_dm(dm_fp, map_fp, filter_cat, filter_vals, keep_vals)
}
\arguments{
\item{dm_fp}{Dissimilarity matrix filepath (tab-delimited text).}

\item{map_fp}{Metadata mapping filepath.}

\item{filter_cat}{[OPTIONAL] The map_fp header string for the factor you
would like to use to filter samples.}

\item{filter_vals}{[OPTIONAL] The values within the filter category (vector
or single value) you would like to use to remove samples from the imported
data.}

\item{keep_vals}{[OPTIONAL] Alternatively, keep only samples represented by
these values.}
}
\value{
A list variable with (1) the loaded dissimilarity matrix, and (2) the
  loaded mapping file.
}
\description{
Load in a dissimilarity matrix and a corresponding metadata 
  mapping file
}
\examples{
\dontrun{
load_dm("filepath_to_dissim_matrix.txt", "filepath_to_mapping_file.txt",
  "sample_type", filter_vals = "blank")
}
}
\concept{
Load external data
}

