% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_output_filter.R
\name{load_taxa_table}
\alias{load_taxa_table}
\title{Load a taxa table for use with mctoolsr}
\usage{
load_taxa_table(tab_fp, map_fp, filter_cat, filter_vals, keep_vals, taxa_fp)
}
\arguments{
\item{tab_fp}{Taxa table filepath.}

\item{map_fp}{Metadata mapping filepath.}

\item{filter_cat}{[OPTIONAL] The map_fp header string for the factor you
would like to use to filter samples.}

\item{filter_vals}{[OPTIONAL] The values within the filter category (vector
or single value) you would like to use to remove samples from the imported
data.}

\item{keep_vals}{[OPTIONAL] Alternatively, keep only samples represented by
these values.}

\item{taxa_fp}{[OPTIONAL] If taxonomy is not contained in the taxa table, you
can supply a tab-delimited file with the taxon ID in the first column and
the taxonomy in the second column. It assumes no header line in the file. 
Lines beginning with '#' are ignored.}
}
\value{
A list variable with (1) the loaded taxa table, (2) the loaded 
  mapping file, and optionally (3) the loaded taxonomy.
}
\description{
Load in a taxa table (also known as an OTU table) and a
  corresponding mapping file with metadata values. The samples in the loaded
  taxa table and mapping file will be in the same order and only samples in
  both will be loaded. The function can optionally filter samples of a
  specific type based on the mapping file. This can also be done later via
  the \code{\link{filter_data}} function.
}
\examples{
\dontrun{
load_taxa_table("filepath_to_taxa_table.txt", "filepath_to_mapping_file.txt",
  "sample_type", filter_vals = "blank")
}
}
\concept{
Load external data
}

