% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/routine_analysis_functions.R
\name{load_taxon_table}
\alias{load_taxon_table}
\title{Load a taxon table for use with mctoolsr}
\usage{
load_taxon_table(tab_fp, map_fp, filter_cat, filter_vals, keep_vals)
}
\arguments{
\item{tab_fp}{Taxon table file path}
}
\value{
A list variable with (1) the loaded taxon table, and (2) the loaded mapping file
}
\description{
Load in a taxon table (aka. an OTU table) and a corresponding mapping file
with metadata values. The samples in the loaded taxon table and mapping file
will be in the same order and only samples in both will be loaded. The function
can optionally filter samples of a specific type based on the mapping file.
This can also be done later via the filter_data() function.
}

