% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree-based.R
\name{load_tree}
\alias{load_tree}
\title{Load in a phylogenetic tree}
\usage{
load_tree(tree_fp)
}
\arguments{
\item{tree_fp}{The phylogenetic tree file path.}
}
\description{
A wrapper for \code{\link[ape]{read.tree}} in
  \code{\link[ape]{ape}}. Reads in a phylogenetic tree in Newic or New
  Hampshire format.
}
\examples{
\dontrun{
load_tree("tree file path")
}
}
\concept{
Load external data
}
