% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mctoolsr.R
\docType{package}
\name{mctoolsr}
\alias{mctoolsr}
\alias{mctoolsr-package}
\title{Microbial Community Data Analysis Tools}
\description{
The mctoolsr package provides various functions for facilitating data
analysis of microbial communities as determined using high throughput
sequenceing. The tools generally start with the taxon table and associated
metadata after the initial processing has been accomplished.
}
\section{Loading data into R using mctoolsr}{

There are multiple functions used to load in data, but the most useful is:
\code{load_taxon_table()}. This function will make sure your community data
and metadata correspond to one another and the samples are in the same order.
}

\section{Microbial community composition functions}{

Need to add more documentation here
}

\section{Taxonomy-based functions}{

Need to add more documentation here
}

\section{Diversity-based functions}{

Need to add more documentation here
}
\concept{
dont include
}

