% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dissimilarity-based.R
\name{plot_dendrogram}
\alias{plot_dendrogram}
\title{Generate a dendrogram based on a dissimilarity matrix}
\usage{
plot_dendrogram(dm, metadata_map, labels, color_by, method = "complete", ...)
}
\arguments{
\item{dm}{Dissimilarity matrix.}

\item{metadata_map}{The metadata mapping dataframe.
Typically, input$map_loaded.}

\item{labels}{The metadata mapping dataframe column name representing the
intended leaf labels.}

\item{color_by}{The metadata mapping dataframe column name representing the
intended leaf label colors.}

\item{method}{The clustering method to use when creating the dendrogram.}

\item{...}{Other parameters passed on to geom_text}
}
\description{
Generate a dendrogram based on a dissimilarity matrix
}
\concept{
Plots
}

