% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/diversity-based.R
\name{plot_diversity}
\alias{plot_diversity}
\title{Plot diversity values}
\usage{
plot_diversity(input, variable, metric)
}
\arguments{
\item{input}{The input dataset as loaded by \code{load_taxa_table()}.}

\item{variable}{Variable in mapping file to plot by.}

\item{metric}{Diversity metric to use.}
}
\description{
Plot diversity using boxplots across levels of a given factor.
}

