% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dissimilarity-based.R
\name{plot_nmds}
\alias{plot_nmds}
\title{Generate an NMDS Plot Quickly}
\usage{
plot_nmds(dm, metadata_map = NULL, color_cat, shape_cat)
}
\arguments{
\item{dm}{Dissimilarity matrix.}

\item{metadata_map}{The metadata mapping data frame.}

\item{color_cat}{The metadata map header used to color points.}

\item{shape_cat}{[OPTIONAL] The metadata map header used for points' shapes.}
}
\description{
Used to generate a quick ordination.
}
\examples{
dm = calc_dm(fruits_veggies$data_loaded)
plot_nmds(dm, metadata_map = fruits_veggies$map_loaded, "Sample_type",
          "Farm_type")
}
\concept{
Plots
}

