% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dissimilarity-based.R
\name{plot_ordination}
\alias{plot_ordination}
\title{Generate an Ordination Plot}
\usage{
plot_ordination(input, ordination_axes, color_cat, shape_cat, hulls = FALSE)
}
\arguments{
\item{input}{The input dataset as loaded by \code{load_taxa_table()}.}

\item{ordination_axes}{The output of \code{calc_ordination()}.}

\item{color_cat}{The metadata map header used to color points.}

\item{shape_cat}{The metadata map header used for points' shapes (optional).}

\item{hulls}{Whether or not to include an outline around sample categories.}
}
\description{
Used to generate a plot from the output of \code{
 calc_ordination()}
}

