% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/taxonomy-based.R
\name{plot_taxa_bars}
\alias{plot_taxa_bars}
\title{Plot stacked bar plots to represent taxa compompositions}
\usage{
plot_taxa_bars(taxa_summary_df, metadata_map, group_factor, num_taxa)
}
\arguments{
\item{taxa_summary_df}{The taxon summary dataframe (the output of \code{
summarize_taxonomy()})}

\item{metadata_map}{The mapping file dataframe}

\item{group_factor}{The factor (metadata header label) used to create the
bars. Means will be taken for each factor level.}

\item{num_taxa}{The number of top most abundant taxa to display. Additional
will be grouped into "Other".}
}
\description{
Stacked bar plots will be generated for each factor level
  indicated by \code{group_factor} to display their taxonomic compositions.
  Only the top few taxa will be displayed as indicated by \code{num_taxa}.
  Mean values are calculated within the factor levels.
}

