% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/misc_plots.R
\name{plot_venn_diagram}
\alias{plot_venn_diagram}
\title{Plot Venn Diagrams}
\usage{
plot_venn_diagram(input, category, pres_thresh)
}
\arguments{
\item{input}{The input dataset as loaded by \code{load_taxa_table()}.}

\item{category}{The category (header in map) to group samples by.}

\item{pres_thresh}{The threshold relative abundance for an OTU to be
considered present for a given sample category.}
}
\description{
Plot venn diagram showing OTU overlap between multiple sample
 categories.
}
\concept{
Plots
}

