% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/normalization.R
\name{single_rarefy}
\alias{single_rarefy}
\title{Rarefy samples in a taxa table}
\usage{
single_rarefy(input, depth)
}
\arguments{
\item{input}{The input dataset as loaded by \code{\link{load_taxa_table}}.}

\item{depth}{The number of observations to randomly select for each sample.}
}
\value{
Will return a dataset list similar to the input.
}
\description{
Rarefying means to randomly select a specified number of
 observations from each sample. This approach is used to normalize for
 differences in the number of observations (i.e. sequencing depth) across
 samples. Samples with fewer observations (sequences) than the specified
 depth will be removed.
}
\examples{
single_rarefy(fruits_veggies, 1000)
}
\concept{
Taxa table normalization
}

