% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets_documentation.R
\docType{data}
\name{sumtax_example_data}
\alias{sumtax_example_data}
\title{Taxa Summary from Produce Dataset}
\format{A data frame}
\source{
Adapted from \url{
 http://journals.plos.org/plosone/article?id=10.1371/journal.pone.0059310}
}
\usage{
sumtax_example_data
}
\description{
Taxa (phyla) relative abundances across all samples in the example dataset.
Produced using \code{mctoolsr::summarize_taxonomy()}.
}
\keyword{datasets}

