% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dissimilarity-based.R
\name{add_metadata_to_dm_clmns}
\alias{add_metadata_to_dm_clmns}
\title{Add metadata to an additional column in column formatted 
  dissimilarities dataframe}
\usage{
add_metadata_to_dm_clmns(dm_clmns, metadata_map, cat)
}
\arguments{
\item{dm_clmns}{The dissimilarities dataframe produced using \code{ 
convert_dm_to_3_column()}.}

\item{metadata_map}{The metadata dataframe.}

\item{cat}{The header string from the metadata map corresponding to the data 
you would like to add to the dissimilarities dataframe.}
}
\value{
A 5 column data frame with the metadata added. The two additional 
  columns correspond to x1 and x2's metadata values
}
\examples{
DM = calc_dm(fruits_veggies$data_loaded)
DM_cols = convert_dm_to_3_column(DM)
DM_cols_with_metadata = add_metadata_to_dm_clmns(DM_cols, 
fruits_veggies$map_loaded, "Sample_type")  
}
\concept{
Dissimilarity calculation and manipulation
}

