% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/diversity-based.R
\name{calc_diversity}
\alias{calc_diversity}
\title{Calculate diversity values}
\usage{
calc_diversity(tax_table, metric)
}
\arguments{
\item{tax_table}{The taxa table.}

\item{metric}{The diversity metric to calculate. Acceptable values are:
'richness', 'shannon', and 'simpson'.}
}
\description{
Calculate diversity values from a taxa table.
}
\details{
Richness is the number of uniaue species per sample. Simpson's
 diversity metric (1 - D) quantifies evenness, and Shannon diversity is a
 combination of richness and evenness.
}

