% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dissimilarity-based.R
\name{calc_mean_dissimilarities}
\alias{calc_mean_dissimilarities}
\title{Calculate mean dissimilarities using a metadata factor}
\usage{
calc_mean_dissimilarities(dm, metadata_map, summarize_by_factor,
  return_map = FALSE)
}
\arguments{
\item{dm}{Dissimilarity matrix - typically created using
\code{\link{calc_dm}}.}

\item{metadata_map}{The metadata mapping dataframe.}

\item{summarize_by_factor}{Category in mapping file to summarize by.}

\item{return_map}{Whether or not to return summarized mapping files. If true,
will return a list (default: FALSE).}
}
\value{
Mean dissimilarities.
}
\description{
Calculate mean dissimilarities across all levels of a given
  factor
}
\examples{
DM = calc_dm(fruits_veggies$data_loaded)
MD = calc_mean_dissimilarities(DM, fruits_veggies$map_loaded, "Sample_type")
}
\concept{
Dissimilarity calculation and manipulation
}

