% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dissimilarity-based.R
\name{calc_ordination}
\alias{calc_ordination}
\title{Calculate Point Coordinates in an Ordination}
\usage{
calc_ordination(dm, ord_type, metadata_map, constrain_factor)
}
\arguments{
\item{dm}{Dissimilarity matrix.}

\item{ord_type}{The type of ordination. 'NMDS' or 'constrained' are the
current accepted values.}

\item{metadata_map}{Required if 'constrained' ord_type.}

\item{constrain_factor}{Required if 'constrained' ord_type.}
}
\description{
Use before plotting ordination.
}
\concept{
Dissimilarity calculation and manipulation
}

