% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_analyses.R
\name{calc_taxa_changes}
\alias{calc_taxa_changes}
\title{Calculate the changes in taxonomic relative abundances compared to 
  controls}
\usage{
calc_taxa_changes(ts, metadata_map, block_header, treatment_header,
  control_label, zero_substitute)
}
\arguments{
\item{ts}{A taxa summary dataframe as generated by 
\code{\link{summarize_taxonomy}}.}

\item{metadata_map}{A metadata mapping dataframe. Usually, 
\code{input$map_loaded}.}

\item{block_header}{The header name from the metadata map that contains the 
block (i.e. sample group) designations.}

\item{treatment_header}{The header name from the metadata map that contains 
the treatment designations including which samples are controls.}

\item{control_label}{The designation within the \code{treatment_header} 
column that corresponds to the control samples.}

\item{zero_subsitute}{[OPTIONAL] A value to substitute in for zeros to avoid
infinite percent changes. It can be appropriate to use a lower detection
limit for this value.}
}
\description{
Calculate the change relative to controls (percent) within 
  sample groups (blocks) on a per taxon basis.
}
\examples{
# calculate the percent difference in organic from conventional for each 
# fruit/veggie
ts = summarize_taxonomy(fruits_veggies, level = 2)
ts_filt = filter_taxa_from_table(ts, filter_thresh = 0.01)
calc_taxa_changes(ts_filt, fruits_veggies$map_loaded, 'Sample_type',
                  'Farm_type', 'Conventional')
}
\concept{
Misc analyses
}

