% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomy-based.R
\name{collapse_taxonomy}
\alias{collapse_taxonomy}
\title{Collapse taxonomy dataframe to character vector}
\usage{
collapse_taxonomy(taxonomy_df)
}
\arguments{
\item{taxonomy_df}{The dataframe containing taxonomy as loaded by
\code{\link{load_taxa_table}}.}
}
\description{
A quick way to get taxonomy strings for each taxon
}
\examples{
collapse_taxonomy(fruits_veggies$taxonomy_loaded)
}
\concept{Taxonomy-based analyses}
