% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dissimilarity-based.R
\name{convert_dm_to_3_column}
\alias{convert_dm_to_3_column}
\title{Convert dissimilarity matrix to 3 column format}
\usage{
convert_dm_to_3_column(dm)
}
\arguments{
\item{dm}{Dissimilarity matrix of either class 'dist' or class 'data.frame'}
}
\value{
The dissimilarity matrix as a 3 column data frame with coulmns: 
  c('x1', 'x2', 'dist')
}
\description{
This is useful for performing analyses on dissimilarity values
}
\examples{
DM = calc_dm(fruits_veggies$data_loaded)
DM_cols = convert_dm_to_3_column(DM) 
}
\concept{
Dissimilarity calculation and manipulation
}

