% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.R
\name{convert_to_relative_abundances}
\alias{convert_to_relative_abundances}
\title{Convert taxa table to relative abundances}
\usage{
convert_to_relative_abundances(input)
}
\arguments{
\item{input}{Either a dataset as generated using
\code{\link{load_taxa_table}} which includes a mapping file or an
individual taxon table}
}
\description{
Convert taxa table or taxon table in data input to relative 
  abundances
}
\examples{
convert_to_relative_abundances(fruits_veggies)
}
\concept{Taxa table normalization}
