% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/input_output_filter.R
\name{export_otu_table}
\alias{export_otu_table}
\title{Export an OTU table as a text file [DEPRECATED]}
\usage{
export_otu_table(input, out_fp, map_fp)
}
\arguments{
\item{input}{The input dataset as loaded by \\code{\link{load_taxa_table}} or
an otu table of class \code{data.frame}.}

\item{out_fp}{The output filepath.}

\item{map_fp}{[OPTIONAL] The metadata map output filepath if you want to
write it to file.}
}
\description{
PLEASE USE \code{export_taxa_table()} INSTEAD.
}
\concept{
dont include
}

