% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/input_output.R
\name{export_otu_table}
\alias{export_otu_table}
\title{Export an OTU table as a text file}
\usage{
export_otu_table(input, out_fp)
}
\arguments{
\item{input}{The input dataset as loaded by \code{load_taxa_table()}}

\item{out_fp}{The output filepath}
}
\description{
A convenient way to export a loaded OTU table as a text file.
 Taxonomy strings will appear in the right most column. This is also a good
 way to save an OTU table to be loaded later.
}

