% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/input_output_filter.R
\name{filter_dm}
\alias{filter_dm}
\title{Filter Samples from Dissimilarity Matrix}
\usage{
filter_dm(input_dm, filter_cat, filter_vals, keep_vals)
}
\arguments{
\item{input_dm}{The input dissimilarity matrix with corresponding mapping
file as generated by \code{load_dm()} or \code{calc_mean_dissimilarities()}
with the option to produce a resulting metadata map.}

\item{filter_cat}{The metadata map header string for the factor you would
like to use to filter samples.}

\item{filter_vals}{The values within the filter category (vector or single
value) you would like to use to remove samples from the dissimilarity
matrix and metadata map.}

\item{keep_vals}{Alternatively, keep only samples represented by these
values.}
}
\value{
A list variable with (1) the loaded dissimilarity matrix and (2)
 the loaded mapping file.
}
\description{
Filter out or keep particular samples in a dissimilarity matrix
 based on contextual metadata.
}

