% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/taxonomy_based.R
\name{filter_taxa_from_data}
\alias{filter_taxa_from_data}
\title{Filter taxa from a loaded dataset}
\usage{
filter_taxa_from_data()
}
\description{
Filter taxa from a loaded dataset
}
\details{
This function is now depricated to keep names consistent. Use
  \code{filter_taxa_from_input()} instead.
}

