% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/taxonomy_based.R
\name{filter_taxa_from_data}
\alias{filter_taxa_from_data}
\title{Filter taxa from a loaded dataset (depricated)}
\usage{
filter_taxa_from_data(input, filter_thresh, taxa_to_keep, taxa_to_remove,
  at_spec_level)
}
\arguments{
\item{input}{Input data (a list variable) from \code{load_taxa_table()}.}

\item{filter_thresh}{Filter OTUs less than this number based on mean OTU
table values.}

\item{taxa_to_keep}{Keep only taxa that contain these names. Vector or string.}

\item{taxa_to_remove}{Remove taxa that contain these names. Vector or string.}

\item{at_spec_level}{If included, only keep/remove matches at this specific
taxonomy level(s) (a number/numbers referring to the taxonomy
column(s)).}
}
\description{
Filter taxa from a loaded dataset (depricated)
}
\details{
This function is now depricated to keep names consistent. Use
 \code{filter_taxa_from_input()} instead.
}

