% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/routine_analysis_functions.R
\name{load_taxon_table}
\alias{load_taxon_table}
\title{Load a taxon table for use with mctoolsr}
\usage{
load_taxon_table(tab_fp, map_fp, filter_cat, filter_vals, keep_vals)
}
\arguments{
\item{tab_fp}{Taxon table filepath.}

\item{map_fp}{Metadata mapping filepath.}

\item{filter_cat}{The map_fp header string for the factor you would like
to use to filter samples.}

\item{filter_vals}{The values within the filter category (vector or single
value) you would like to use to remove samples from the imported data.}

\item{keep_vals}{Alternatively, keep only samples represented by these
values.}
}
\value{
A list variable with (1) the loaded taxon table, and (2) the loaded
 mapping file.
}
\description{
Load in a taxon table (aka. an OTU table) and a corresponding
 mapping file with metadata values. The samples in the loaded taxon table
 and mapping file will be in the same order and only samples in both will
 be loaded. The function can optionally filter samples of a specific type
 based on the mapping file. This can also be done later via the filter_data()
 function.
}
\examples{
\dontrun{
load_taxon_table("filepath_to_OTU_table.txt", "filepath_to_mapping_file.txt",
  "sample_type", filter_vals = "blank")
}
}

