% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tree-based.R
\name{load_tree}
\alias{load_tree}
\title{Load in a phylogenetic tree}
\usage{
load_tree(tree_fp)
}
\arguments{
\item{tree_fp}{The phylogenetic tree file path.}
}
\description{
A wrapper for \code{read.tree()} in \code{ape}. Reads in a
 phylogenetic tree in Newic or New Hampshire format.
}

