% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/input_output.R
\name{match_datasets}
\alias{match_datasets}
\title{Match up samples from two datasets}
\usage{
match_datasets(ds1, ds2)
}
\arguments{
\item{ds1,}{ds2 The two datasets as loaded by \code{load_taxa_table()}}
}
\value{
A list variable with the matched ds1 as the first element and ds2
 as the second element
}
\description{
Function to match up sample order from two datasets that contain
 some overlapping sample IDs. Sample IDs that are not present in both
 datasets will be dropped. The output is a list containing the two filtered
 datasets in the same order as they were input.
}

