% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dissimilarity-based.R
\name{plot_dendrogram}
\alias{plot_dendrogram}
\title{Generate a dendrogram based on a dissimilarity matrix}
\usage{
plot_dendrogram(dm, metadata_map, labels, color_by, method = "complete", ...)
}
\arguments{
\item{dm}{Dissimilarity matrix.}

\item{metadata_map}{The metadata mapping dataframe. Typically, 
input$map_loaded.}

\item{labels}{The metadata mapping dataframe column name representing the 
intended leaf labels.}

\item{color_by}{[OPTIONAL] The metadata mapping dataframe column name 
representing the intended leaf label colors.}

\item{method}{The clustering method to use when creating the dendrogram.}

\item{...}{Other parameters passed on to geom_text}
}
\description{
Generating a dendrogram can be useful for visualizing 
  similarities and differences in community compositions among samples.
}
\examples{
fvrar = single_rarefy(fruits_veggies, 100)
dm = calc_dm(fvrar$data_loaded)
plot_dendrogram(
  dm, metadata_map = fvrar$map_loaded, labels = 'Sample_type',
  color_by = 'Farm_type'
)
}
\concept{Plots}
