% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/routine_analysis_functions.R
\name{plot_diversity}
\alias{plot_diversity}
\title{Plot diversity values}
\usage{
plot_diversity(input, variable, metric)
}
\arguments{
\item{variable}{Variable in mapping file to plot by}
}
\description{
Plot diversity using boxplots across levels of a given factor
}

