% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dissimilarity-based.R
\name{plot_nmds}
\alias{plot_nmds}
\title{Generate an NMDS Plot Quickly}
\usage{
plot_nmds(dm, metadata_map = NULL, color_cat, shape_cat)
}
\arguments{
\item{dm}{Dissimilarity matrix.}

\item{metadata_map}{The metadata mapping dataframe.}

\item{color_cat}{The metadata map header used to color points.}

\item{shape_cat}{The metadata map header used for points' shapes (optional).}
}
\description{
Used to generate a quick ordination.
}

