% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dissimilarity-based.R
\name{plot_ordination}
\alias{plot_ordination}
\title{Generate an Ordination Plot}
\usage{
plot_ordination(input, ordination_axes, color_cat, shape_cat, hulls = FALSE,
  ...)
}
\arguments{
\item{input}{The input dataset as loaded by \code{load_taxa_table()}.}

\item{ordination_axes}{The output of \code{calc_ordination()}.}

\item{color_cat}{The metadata map header used to color points.}

\item{shape_cat}{The metadata map header used for points' shapes (optional).}

\item{hulls}{Whether or not to include an outline around sample categories.}

\item{...}{Additional arguments passed to ggplot2's \code{geom_point()}.}
}
\description{
Used to generate a plot from the output of \code{ 
  calc_ordination()}
}
\examples{
fvrar = single_rarefy(fruits_veggies, 100)
ord = calc_ordination(calc_dm(fvrar$data_loaded), 'nmds')
plot_ordination(fvrar, ord, 'Sample_type', 'Farm_type', hulls = TRUE)
}
\concept{
Plots
}

