% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/taxonomy-based.R
\name{plot_taxa_bars}
\alias{plot_taxa_bars}
\title{Plot stacked bar plots to represent taxa compompositions}
\usage{
plot_taxa_bars(tax_table, metadata_map, type_header, num_taxa,
  data_only = FALSE)
}
\arguments{
\item{tax_table}{A taxa table dataframe. The output of
\code{\link{summarize_taxonomy}}.}

\item{metadata_map}{The mapping file dataframe.}

\item{type_header}{The factor (metadata header label) used to create the
bars. Means will be taken for each factor level.}

\item{num_taxa}{The number of top most abundant taxa to display. Additional
will be grouped into "Other".}

\item{data_only}{[OPTIONAL] Set to \code{TRUE} if you want the plotting data
returned instead of the plot.}
}
\description{
Stacked bar plots will be generated for each factor level
  indicated by \code{type_header} to display their taxonomic compositions.
  Only the top few taxa will be displayed as indicated by \code{num_taxa}.
  Mean values are calculated within the factor levels.
}
\concept{
Plots
}

