% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/taxonomy_based.R
\name{plot_ts_heatmap}
\alias{plot_ts_heatmap}
\title{Plot Taxa Summary Heatmap}
\usage{
plot_ts_heatmap(tax_table, metadata_map, min_rel_abund, type_header, scale_by)
}
\arguments{
\item{tax_table}{A taxo table dataframe.}

\item{metadata_map}{A metadata mapping dataframe.}

\item{min_rel_abund}{The minimum mean relative abundance for a taxon to not
be grouped into 'Other'.}

\item{type_header}{The metadata_map header label used to group samples.}

\item{scale_by}{Whether to scale colors by (a) 'sample_types', (b) 'taxa', or
(c) 'all'.}
}
\description{
A quick way to create a heatmap from a taxa summary dataframe.
 Samples are grouped by a category that is specified in 'type_header'.
}

