% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/taxonomy-based.R
\name{plot_ts_heatmap}
\alias{plot_ts_heatmap}
\title{Plot Taxa Summary Heatmap}
\usage{
plot_ts_heatmap(tax_table, metadata_map, min_rel_abund, type_header, scale_by,
  custom_sample_order)
}
\arguments{
\item{tax_table}{A taxa table dataframe. The output of
\code{summarize_taxonomy()}.}

\item{metadata_map}{A metadata mapping dataframe. Usually,
\code{input$map_loaded}}

\item{min_rel_abund}{The minimum mean relative abundance for a taxon to not
be grouped into 'Other'. Between 0 and 1.}

\item{type_header}{The metadata_map header label used to group samples.}

\item{scale_by}{Whether to scale colors by (a) 'sample_types', (b) 'taxa', or
(c) 'all'.}

\item{custom_sample_order}{An optional vector with the order of the sample
names (top to bottom).}
}
\description{
A quick way to create a heatmap from a taxa summary dataframe.
 Samples are grouped by a category that is specified in 'type_header'.
}

